requirements.txt
fastapi
uvicorn[standard]
transformers
torch
opencv-python
Pillow
sqlalchemy
psycopg2-binary
easyocr
python-dotenv